+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 
 # Unreleased
+- `Project.get_jobs_status()` now accepts `**kwargs` to select which jobs to check ([#662](https://github.com/pyiron/pyiron_base/pull/662))
 
 # pyiron_base-0.5.4
 - Add functionality to 'defragment' the hdf storage within a project (local maintenance). If datasets in the hdf are overwritten, the old memory allocation is never used again and the hdf file increses in size. Rewriting it reduces the size again, which is now available at `Project('path').maintenance.local.defragment_storage()` ([#673](https://github.com/pyiron/pyiron_base/pull/673))